added utility for parsing whole hdf5. patched sequencer into current dataloader.
